Backend/requirements.txt
flask
flask-pymongo
flask-jwt-extended
flask-cors
bcrypt
python-dotenv
requests
cloudinary
pymongo
